refactor: Canvas 処理を切り出した
